requirements.txt
Django==4.2.11
djangorestframework==3.15.1
pre-commit==3.7.0
coverage==7.4.4
django-cors-headers==4.3.1
psycopg2-binary==2.9.6
requests
